+++ crates/canic/src/macros/endpoints/root.rs
@@ -46,13 +46,13 @@ macro_rules! canic_emit_root_admin_endpoints {
             Ok($crate::__internal::core::api::topology::registry::AppRegistryApi::registry())
         }
 
-        #[$crate::canic_query]
+        #[$crate::canic_query(public)]
         fn canic_subnet_registry()
         -> Result<::canic::dto::topology::SubnetRegistryResponse, ::canic::Error> {
             Ok($crate::__internal::core::api::topology::registry::SubnetRegistryApi::registry())
         }
 
-        #[$crate::canic_query]
+        #[$crate::canic_query(public)]
         async fn canic_pool_list()
         -> Result<::canic::dto::pool::CanisterPoolResponse, ::canic::Error> {
             Ok($crate::__internal::core::api::pool::CanisterPoolApi::list())